requirements.txt

torch
torchvision
streamlit
numpy
pandas
pillow
matplotlib
scikit-learn
watchdog
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
requests
extra-streamlit-components
opencv-python-headless
